services/orchestrator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.3.0
